Issue-100: Fix async Filters double-wrapping exceptions
closes #100
SemVer:Patch
